Remove chat only by create, show only chat for you

+++ chatroom.dev/chat_delete.php
@@ -8,27 +8,41 @@
 
  $name = $_POST['chatname'];
  $name = pg_escape_string($name);
- 
- $sql = "DELETE FROM chatrooms WHERE chatname='{$name}'";
- 
- $ret = pg_query($link, $sql);
 
- if (!$ret) {
-  echo pg_last_error($link);
- } else {
-  echo "Chatroom removed from chatrooms </br>";
- }
-
- $sql = "DROP TABLE IF EXISTS $name";
+ // Only creator can delete chatroom 
+ $sql = "SELECT * FROM chatrooms WHERE chatname='{$name}'";
  
  $ret = pg_query($link, $sql);
- if (!$ret) {
+
+ if (!$ret){
   echo pg_last_error($link);
  } else {
-  echo "Table deleted successfully </br>";
- }
- echo "Name: " . $name;
+  $row = pg_fetch_assoc($ret);
+  $logged = trim($_SESSION['login_user']);
+  $creator = trim($row['creator']);
+  if (strcmp($creator, $logged) == 0){   
 
+   $sql = "DELETE FROM chatrooms WHERE chatname='{$name}'";
+ 
+   $ret = pg_query($link, $sql);
+
+   if (!$ret) {
+    echo pg_last_error($link);
+   } else {
+    echo "Chatroom removed from chatrooms </br>";
+   }
+
+   $sql = "DROP TABLE IF EXISTS $name";
+ 
+   $ret = pg_query($link, $sql);
+   if (!$ret) {
+    echo pg_last_error($link);
+   } else {
+    echo "Table deleted successfully </br>";
+   }
+  }
+ }
+ 
  // close connection to db 
  pg_close($link);
